# docs/swagger.yaml
definitions:
  api.Alert:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      endsAt:
        type: string
      fingerprint:
        type: string
      generatorURL:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      startsAt:
        type: string
      status:
        type: string
    type: object
  api.CreateNewChat:
    properties:
      language:
        example: en
        type: string
      message:
        example: Pod web-server-5b866987d8-sxmtj in namespace default Crashlooping.
        type: string
    type: object
  api.CreateNewKnowledge:
    properties:
      gitAuthToken:
        example: ""
        type: string
      gitBranch:
        example: ""
        type: string
      gitRepositoryURL:
        example: ""
        type: string
      path:
        example: ./docs/runbooks
        type: string
      source:
        example: local
        type: string
    type: object
  api.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.FetchMessagesResponse:
    properties:
      chat:
        type: string
      response:
        type: string
    type: object
  api.PurgeKnowledge:
    type: object
  api.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  api.WebhookPayload:
    properties:
      alerts:
        items:
          $ref: '#/definitions/api.Alert'
        type: array
      commonAnnotations:
        additionalProperties:
          type: string
        type: object
      commonLabels:
        additionalProperties:
          type: string
        type: object
      externalURL:
        type: string
      groupKey:
        type: string
      groupLabels:
        additionalProperties:
          type: string
        type: object
      receiver:
        type: string
      status:
        type: string
      truncatedAlerts:
        type: integer
      version:
        type: string
    type: object
  slack.ReceiveSlackEvent:
    properties:
      api_app_id:
        type: string
      authorizations:
        items:
          $ref: '#/definitions/slack.SlackAuthorization'
        type: array
      challenge:
        type: string
      context_enterprise_id:
        type: string
      context_team_id:
        type: string
      event:
        $ref: '#/definitions/slack.SlackEvent'
      event_context:
        type: string
      event_id:
        type: string
      event_time:
        type: integer
      is_ext_shared_channel:
        type: boolean
      team_id:
        type: string
      token:
        type: string
      type:
        type: string
    type: object
  slack.SlackAttachment:
    properties:
      color:
        type: string
      fallback:
        type: string
      id:
        type: integer
      mrkdwn_in:
        items:
          type: string
        type: array
      text:
        type: string
      title:
        type: string
      title_link:
        type: string
    type: object
  slack.SlackAuthorization:
    properties:
      enterprise_id:
        type: string
      is_bot:
        type: boolean
      is_enterprise_install:
        type: boolean
      team_id:
        type: string
      user_id:
        type: string
    type: object
  slack.SlackBlock:
    properties:
      block_id:
        type: string
      elements:
        items:
          $ref: '#/definitions/slack.SlackBlockSection'
        type: array
      type:
        type: string
    type: object
  slack.SlackBlockElement:
    properties:
      text:
        type: string
      type:
        type: string
    type: object
  slack.SlackBlockSection:
    properties:
      elements:
        items:
          $ref: '#/definitions/slack.SlackBlockElement'
        type: array
      type:
        type: string
    type: object
  slack.SlackEvent:
    properties:
      attachments:
        items:
          $ref: '#/definitions/slack.SlackAttachment'
        type: array
      blocks:
        items:
          $ref: '#/definitions/slack.SlackBlock'
        type: array
      bot_id:
        type: string
      channel:
        type: string
      channel_type:
        type: string
      client_msg_id:
        type: string
      event_ts:
        type: string
      subtype:
        type: string
      team:
        type: string
      text:
        type: string
      thread_ts:
        type: string
      ts:
        type: string
      type:
        type: string
      user:
        type: string
    type: object
info:
  contact: {}
paths:
  /chat:
    post:
      consumes:
      - application/json
      description: used to chat with Ava
      parameters:
      - description: Create a new chat
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/api.CreateNewChat'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Chat with Ava
      tags:
      - Chat
  /chat/{id}:
    get:
      consumes:
      - application/json
      description: used to chat with Ava
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FetchMessagesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Chat with Ava
      tags:
      - Chat
    post:
      consumes:
      - application/json
      description: used to respond to Ava
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Create a new chat
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/api.CreateNewChat'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Chat with Ava
      tags:
      - Chat
  /chat/webhook:
    post:
      consumes:
      - application/json
      description: used to chat with Ava from AlertManager Webhook
      parameters:
      - description: Webhook payload
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/api.WebhookPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Chat with Ava from AlertManager Webhook
      tags:
      - Chat
  /event/slack:
    post:
      consumes:
      - application/json
      description: used to chat with Ava when a slack event is received
      parameters:
      - description: ReceiveSlackEvent payload
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/slack.ReceiveSlackEvent'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Receive a Slack event and chat with Ava
      tags:
      - Event
  /knowledge:
    delete:
      consumes:
      - application/json
      description: used to purge Ava's knowledge base
      parameters:
      - description: PurgeKnowledge payload
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/api.PurgeKnowledge'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Purge Ava's knowledge base
      tags:
      - Knowledge
    post:
      consumes:
      - application/json
      description: used to add knowledge to Ava
      parameters:
      - description: CreateNewKnowledge payload
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/api.CreateNewKnowledge'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Add knowledge to Ava
      tags:
      - Knowledge
  /live:
    get:
      consumes:
      - application/json
      description: used by Kubernetes liveness probe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "503":
          description: KO
          schema:
            type: string
      summary: Liveness check
      tags:
      - Kubernetes
  /readyz:
    get:
      consumes:
      - application/json
      description: used by Kubernetes readiness probe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "503":
          description: KO
          schema:
            type: string
      summary: Readiness check
      tags:
      - Kubernetes
swagger: "2.0"